translationsv2 feature flag re introduced

+++ app/api/i18n/v2_support.ts
@@ -1,5 +1,5 @@
 import { ResultSet } from 'api/common.v2/contracts/ResultSet';
-import { getClient } from 'api/common.v2/database/getConnectionForCurrentTenant';
+import { getClient, getConnection } from 'api/common.v2/database/getConnectionForCurrentTenant';
 import { MongoTransactionManager } from 'api/common.v2/database/MongoTransactionManager';
 import { DefaultTranslationsDataSource } from 'api/i18n.v2/database/data_source_defaults';
 import { Translation } from 'api/i18n.v2/model/Translation';
@@ -13,9 +13,23 @@ import { UpsertTranslationsService } from 'api/i18n.v2/services/UpsertTranslatio
 import { ValidateTranslationsService } from 'api/i18n.v2/services/ValidateTranslationsService';
 import { EnforcedWithId } from 'api/odm';
 import { DefaultSettingsDataSource } from 'api/settings.v2/database/data_source_defaults';
+import { tenants } from 'api/tenants';
+import { Db } from 'mongodb';
 import { TranslationContext, TranslationType, TranslationValue } from 'shared/translationType';
 import { LanguageISO6391 } from 'shared/types/commonTypes';
 import { IndexedContextValues } from './translations';
+import migration from '../i18n.v2/migrations';
+
+const cleanUpV2Collections = async (db: Db) => {
+  try {
+    await db.collection('translations_v2').drop({});
+    await db.collection('translations_v2_helper').drop();
+  } catch (e) {
+    if (e.message !== 'ns not found') {
+      throw e;
+    }
+  }
+};
 
 const flattenTranslations = (translation: TranslationType): CreateTranslationsData[] => {
   if (translation.contexts?.length) {
@@ -36,7 +50,7 @@ const flattenTranslations = (translation: TranslationType): CreateTranslationsDa
   return [];
 };
 
-const resultsToV1TranslationType = async (
+export const resultsToV1TranslationType = async (
   tranlationsResult: ResultSet<Translation>,
   onlyLanguage?: LanguageISO6391
 ) => {
@@ -95,49 +109,57 @@ const resultsToV1TranslationType = async (
 };
 
 export const createTranslationsV2 = async (translation: TranslationType) => {
-  const transactionManager = new MongoTransactionManager(getClient());
-  await new CreateTranslationsService(
-    DefaultTranslationsDataSource(transactionManager),
-    new ValidateTranslationsService(
+  if (tenants.current().featureFlags?.translationsV2) {
+    const transactionManager = new MongoTransactionManager(getClient());
+    await new CreateTranslationsService(
       DefaultTranslationsDataSource(transactionManager),
-      DefaultSettingsDataSource(transactionManager)
-    ),
-    transactionManager
-  ).create(flattenTranslations(translation));
+      new ValidateTranslationsService(
+        DefaultTranslationsDataSource(transactionManager),
+        DefaultSettingsDataSource(transactionManager)
+      ),
+      transactionManager
+    ).create(flattenTranslations(translation));
+  }
 };
 
 export const upsertTranslationsV2 = async (translations: TranslationType[]) => {
-  const transactionManager = new MongoTransactionManager(getClient());
-  await new UpsertTranslationsService(
-    DefaultTranslationsDataSource(transactionManager),
-    DefaultSettingsDataSource(transactionManager),
-    new ValidateTranslationsService(
+  if (tenants.current().featureFlags?.translationsV2) {
+    const transactionManager = new MongoTransactionManager(getClient());
+    await new UpsertTranslationsService(
       DefaultTranslationsDataSource(transactionManager),
-      DefaultSettingsDataSource(transactionManager)
-    ),
-    transactionManager
-  ).upsert(
-    translations.reduce<CreateTranslationsData[]>(
-      (flattened, t) => flattened.concat(flattenTranslations(t)),
-      []
-    )
-  );
+      DefaultSettingsDataSource(transactionManager),
+      new ValidateTranslationsService(
+        DefaultTranslationsDataSource(transactionManager),
+        DefaultSettingsDataSource(transactionManager)
+      ),
+      transactionManager
+    ).upsert(
+      translations.reduce<CreateTranslationsData[]>(
+        (flattened, t) => flattened.concat(flattenTranslations(t)),
+        []
+      )
+    );
+  }
 };
 
 export const deleteTranslationsByContextIdV2 = async (contextId: string) => {
-  const transactionManager = new MongoTransactionManager(getClient());
-  await new DeleteTranslationsService(
-    DefaultTranslationsDataSource(transactionManager),
-    transactionManager
-  ).deleteByContextId(contextId);
+  if (tenants.current().featureFlags?.translationsV2) {
+    const transactionManager = new MongoTransactionManager(getClient());
+    await new DeleteTranslationsService(
+      DefaultTranslationsDataSource(transactionManager),
+      transactionManager
+    ).deleteByContextId(contextId);
+  }
 };
 
 export const deleteTranslationsByLanguageV2 = async (language: string) => {
-  const transactionManager = new MongoTransactionManager(getClient());
-  return new DeleteTranslationsService(
-    DefaultTranslationsDataSource(transactionManager),
-    transactionManager
-  ).deleteByLanguage(language);
+  if (tenants.current().featureFlags?.translationsV2) {
+    const transactionManager = new MongoTransactionManager(getClient());
+    return new DeleteTranslationsService(
+      DefaultTranslationsDataSource(transactionManager),
+      transactionManager
+    ).deleteByLanguage(language);
+  }
 };
 
 export const getTranslationsV2ByContext = async (context: string) =>
@@ -147,13 +169,16 @@ export const getTranslationsV2ByContext = async (context: string) =>
     ).getByContext(context)
   );
 
-export const getTranslationsV2ByLanguage = async (language: LanguageISO6391) =>
-  resultsToV1TranslationType(
-    new GetTranslationsService(
-      DefaultTranslationsDataSource(new MongoTransactionManager(getClient()))
-    ).getByLanguage(language),
-    language
-  );
+export const getTranslationsV2ByLanguage = async (language: LanguageISO6391) => {
+  if (tenants.current().featureFlags?.translationsV2) {
+    return resultsToV1TranslationType(
+      new GetTranslationsService(
+        DefaultTranslationsDataSource(new MongoTransactionManager(getClient()))
+      ).getByLanguage(language),
+      language
+    );
+  }
+};
 
 export const getTranslationsV2 = async () =>
   resultsToV1TranslationType(
@@ -168,14 +193,67 @@ export const updateContextV2 = async (
   keysToDelete: string[],
   valueChanges: IndexedContextValues
 ) => {
-  const transactionManager = new MongoTransactionManager(getClient());
-  await new UpsertTranslationsService(
-    DefaultTranslationsDataSource(transactionManager),
-    DefaultSettingsDataSource(transactionManager),
-    new ValidateTranslationsService(
+  if (tenants.current().featureFlags?.translationsV2) {
+    const transactionManager = new MongoTransactionManager(getClient());
+    await new UpsertTranslationsService(
       DefaultTranslationsDataSource(transactionManager),
-      DefaultSettingsDataSource(transactionManager)
-    ),
-    transactionManager
-  ).updateContext(context, keyNamesChanges, valueChanges, keysToDelete);
+      DefaultSettingsDataSource(transactionManager),
+      new ValidateTranslationsService(
+        DefaultTranslationsDataSource(transactionManager),
+        DefaultSettingsDataSource(transactionManager)
+      ),
+      transactionManager
+    ).updateContext(context, keyNamesChanges, valueChanges, keysToDelete);
+  }
+};
+
+export const addLanguageV2 = async (
+  newLanguage: LanguageISO6391,
+  defaultLanguage: LanguageISO6391
+) => {
+  if (tenants.current().featureFlags?.translationsV2) {
+    const [defaultTranslation] = (await getTranslationsV2ByLanguage(defaultLanguage)) || [];
+    const newLanguageTranslations = {
+      ...defaultTranslation,
+      locale: newLanguage,
+      contexts: (defaultTranslation.contexts || []).map(({ _id, ...context }) => context),
+    };
+
+    await createTranslationsV2(newLanguageTranslations);
+    const [result] = (await getTranslationsV2ByLanguage(newLanguage)) || [];
+    return result;
+  }
+};
+
+export const migrateTranslationsToV2 = async () => {
+  const db = getConnection();
+
+  if (!tenants.current().featureFlags?.translationsV2) {
+    await cleanUpV2Collections(db);
+    return false;
+  }
+
+  try {
+    await db.collection('translations_v2_helper').createIndex({ migration_helper: 1 });
+    await db.collection('translations_v2_helper').insertOne({ migration_helper: true });
+  } catch (e) {}
+
+  const needsMigration = await db
+    .collection('translations_v2_helper')
+    .findOneAndUpdate({ migration_helper: true }, { $set: { migrating: true } });
+
+  if (needsMigration.value?.migrating) {
+    return false;
+  }
+
+  if (needsMigration.value?.migrated) {
+    return true;
+  }
+
+  await migration.up(db);
+
+  await db
+    .collection('translations_v2_helper')
+    .findOneAndUpdate({ migration_helper: true }, { $set: { migrated: true, migrating: false } });
+  return false;
 };
